+++ packages/stryker-jest-runner/test/integration/JestConfigEditorSpec.ts
@@ -29,7 +29,7 @@ describe('Integration test for Jest ConfigEditor', () => {
       bail: false,
       collectCoverage: false,
       collectCoverageFrom: [
-        'src/**/*.{js,jsx,mjs}'
+        'src/**/*.{js,jsx}'
       ],
       moduleFileExtensions: [
         'web.js',
@@ -37,45 +37,36 @@ describe('Integration test for Jest ConfigEditor', () => {
         'json',
         'web.jsx',
         'jsx',
-        'node',
-        'mjs'
+        'node'
       ],
       moduleNameMapper: {
+        '^.+\\.module\\.(css|sass|scss)$': 'identity-obj-proxy',
         '^react-native$': 'react-native-web'
       },
       rootDir: projectRoot,
-      setupFiles: [path.join(projectRoot, 'node_modules', 'react-scripts', 'config', 'polyfills.js')],
+      setupFiles: [path.join(projectRoot, 'node_modules', 'react-app-polyfill', 'jsdom.js')],
       setupTestFrameworkScriptFile: undefined,
       testEnvironment: 'jsdom',
       testMatch: [
-        '<rootDir>/src/**/__tests__/**/*.{js,jsx,mjs}',
-        '<rootDir>/src/**/?(*.)(spec|test).{js,jsx,mjs}'
+        '<rootDir>/src/**/__tests__/**/*.{js,jsx}',
+        '<rootDir>/src/**/?(*.)(spec|test).{js,jsx}'
       ],
       testResultsProcessor: undefined,
       testURL: 'http://localhost',
       transform: {
-        '^(?!.*\\.(js|jsx|mjs|css|json)$)': path.join(projectRoot, 'node_modules', 'react-scripts', 'config', 'jest', 'fileTransform.js'),
-        '^.+\\.(js|jsx|mjs)$': path.join(projectRoot, 'node_modules', 'react-scripts', 'config', 'jest', 'babelTransform.js'),
+        '^(?!.*\\.(js|jsx|css|json)$)': path.join(projectRoot, 'node_modules', 'react-scripts', 'config', 'jest', 'fileTransform.js'),
+        '^.+\\.(js|jsx)$': path.join(projectRoot, 'node_modules', 'react-scripts', 'config', 'jest', 'babelTransform.js'),
         '^.+\\\.css$': path.join(projectRoot, 'node_modules', 'react-scripts', 'config', 'jest', 'cssTransform.js'),
       },
       transformIgnorePatterns: [
-        '[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$'
+        '[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$',
+        '^.+\\.module\\.(css|sass|scss)$'
       ],
       verbose: false
     };
 
     assertJestConfig(expectedResult, config.jest.config);
   });
-
-  function assertJestConfig(expected: any, actual: any) {
-    Object.keys(expected).forEach(key => {
-      if (Array.isArray(expected[key])) {
-        expected[key].sort();
-        actual[key].sort();
-      }
-      expect(expected[key]).deep.eq(actual[key]);
-    });
-  }
 
   it('should create a Jest configuration for a React + TypeScript project', () => {
     config.set({ jest: { projectType: 'react-ts' } });
@@ -191,4 +182,14 @@ describe('Integration test for Jest ConfigEditor', () => {
 
     expect(() => jestConfigEditor.edit(config)).to.throw(Error, `No configLoader available for ${projectType}`);
   });
+
+  function assertJestConfig(expected: any, actual: any) {
+    Object.keys(expected).forEach(key => {
+      if (Array.isArray(expected[key])) {
+        expected[key].sort();
+        actual[key].sort();
+      }
+      expect(actual[key]).deep.eq(expected[key]);
+    });
+  }
 });