+++ packages/stryker/src/reporters/ProgressKeeper.ts
@@ -1,17 +1,19 @@
 import { MatchedMutant, Reporter, MutantResult } from 'stryker-api/report';
 import { MutantStatus } from 'stryker-api/report';
+import Timer from '../utils/Timer';
 
 abstract class ProgressKeeper implements Reporter {
-
+  private timer: Timer;
   protected progress = {
     survived: 0,
     tested: 0,
-    total: 0
+    total: 0,
   };
 
   private mutantIdsWithoutCoverage: string[];
 
   public onAllMutantsMatchedWithTests(matchedMutants: ReadonlyArray<MatchedMutant>): void {
+    this.timer = new Timer();
     this.mutantIdsWithoutCoverage = matchedMutants.filter(m => m.scopedTestIds.length === 0).map(m => m.id);
     this.progress.total = matchedMutants.length - this.mutantIdsWithoutCoverage.length;
   }
@@ -23,6 +25,34 @@ abstract class ProgressKeeper implements Reporter {
     if (result.status === MutantStatus.Survived) {
       this.progress.survived++;
     }
+  }
+
+  protected getEtc() {
+    const totalSecondsLeft = Math.floor(this.timer.elapsedSeconds() / this.progress.tested * (this.progress.total - this.progress.tested));
+
+    if (isFinite(totalSecondsLeft) && totalSecondsLeft > 0) {
+      const hours = Math.floor(totalSecondsLeft / 3600);
+      const minutes = Math.floor(totalSecondsLeft / 60 % 60);
+      const seconds = Math.floor(totalSecondsLeft % 60);
+
+      return this.formatEtc(hours, minutes, seconds);
+    } else {
+      return 'n/a';
+    }
+  }
+
+  private formatEtc(hours: number, minutes: number, seconds: number) {
+    let output;
+
+    if (hours > 0) {
+      output = `${hours}h, ${minutes}m, ${seconds}s`;
+    } else if (minutes > 0) {
+      output = `${minutes}m, ${seconds}s`;
+    } else {
+      output = `${seconds}s`;
+    }
+
+    return output;
   }
 }
 export default ProgressKeeper;