+++ packages/grunt-stryker/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.13.11"></a>
+## [0.13.11](https://github.com/stryker-mutator/stryker/compare/grunt-stryker@0.13.9...grunt-stryker@0.13.11) (2018-10-15)
+
+
+### Bug Fixes
+
+* **version:** Version bump for failed release ([8cf9e87](https://github.com/stryker-mutator/stryker/commit/8cf9e87))
+
+
+
+
 <a name="0.13.9"></a>
 ## [0.13.9](https://github.com/stryker-mutator/stryker/compare/grunt-stryker@0.13.8...grunt-stryker@0.13.9) (2018-10-03)