+++ packages/stryker-html-reporter/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.5"></a>
+## [0.16.5](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.16.4...stryker-html-reporter@0.16.5) (2018-10-15)
+
+
+### Bug Fixes
+
+* **any-promise:** Don't register a promise implementation ([#1180](https://github.com/stryker-mutator/stryker/issues/1180)) ([1d3e2f6](https://github.com/stryker-mutator/stryker/commit/1d3e2f6))
+
+
+
+
 <a name="0.16.4"></a>
 ## [0.16.4](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.16.3...stryker-html-reporter@0.16.4) (2018-10-03)